backend/pip_requirements.txt
langchain==0.3.0
fastapi
uvicorn
langgraph
langchain-community
langchain-openai
langsmith
langserve
openai
supabase
pandas
numpy
torch
chronos-forecasting
python-multipart
pydantic
pydantic-settings
python-dotenv
psycopg2-binary
scikit-learn
httpx
aiofiles